Fixes for 64-bit systems (orig by Darren Salt).

+++ edge/src/g_game.cc
@@ -107,7 +107,7 @@ int deathmatch;
 
 int gametic;
 
-long random_seed;
+int random_seed;
 
 // if true, load all graphics at start 
 bool precache = true;
